travel_agent/requirements.txt
google-generativeai
crawl4ai
requests
beautifulsoup4
jinja2
python-dateutil
aiohttp
asyncio
pytz
pillow
